feat: 重构 Application 模块架构优先级
